Move raw conversions into its own module
